requirements.txt
loguru==0.5.3
pyTelegramBotAPI==4.13.0
python-decouple==3.6
python-dotenv==0.20.0
requests==2.31.0
dnspython==2.2.1
Flask==2.2.5
gunicorn==20.1.0
beautifulsoup4==4.12.2